# optimize sets