--- a/fputil/request.go
+++ b/fputil/request.go
@@ -146,10 +146,10 @@ type VersionSignature struct {
 // An IntSignature is a signature on a list of integers.
 type IntSignature struct {
 	OrderedList IntList
-	RequiredSet IntSet
-	OptionalSet IntSet
-	UnlikelySet IntSet
-	ExcludedSet IntSet
+	RequiredSet *IntSet
+	OptionalSet *IntSet
+	UnlikelySet *IntSet
+	ExcludedSet *IntSet
 }
 
 // A StringSignature is a signature on a list of strings.
@@ -292,10 +292,10 @@ func NewStringSignature(s string) (StringSignature, error) {
 // Parse an int signature from a string and return an error on failure.
 func (a *IntSignature) Parse(s string) error {
 	a.OrderedList = IntList{}
-	a.ExcludedSet = make(IntSet)
-	a.UnlikelySet = make(IntSet)
-	a.OptionalSet = make(IntSet)
-	a.RequiredSet = make(IntSet)
+	a.ExcludedSet = new(IntSet)
+	a.UnlikelySet = new(IntSet)
+	a.OptionalSet = new(IntSet)
+	a.RequiredSet = new(IntSet)
 	if len(s) == 0 {
 		return nil
 	}
@@ -321,20 +321,21 @@ func (a *IntSignature) Parse(s string) error {
 		case flagOptional, flagUnlikely, flagExcluded:
 			v = v[1:]
 		}
-		elem, err := strconv.ParseUint(v, 16, 16)
+		elem64bit, err := strconv.ParseUint(v, 16, 16)
+		elem := int(elem64bit)
 		if err != nil {
 			return err
 		}
 		switch flag {
 		case flagOptional:
-			a.OptionalSet[elem] = true
+			a.OptionalSet.Insert(elem)
 		case flagUnlikely:
-			a.UnlikelySet[elem] = true
+			a.UnlikelySet.Insert(elem)
 		case flagExcluded:
-			a.ExcludedSet[elem] = true
+			a.ExcludedSet.Insert(elem)
 			continue // do not add to ordered list
 		default:
-			a.RequiredSet[elem] = true
+			a.RequiredSet.Insert(elem)
 		}
 		a.OrderedList = append(a.OrderedList, elem)
 	}
@@ -342,7 +343,7 @@ func (a *IntSignature) Parse(s string) error {
 		// allow any order and any optional items
 		// still check for required, unlikely, and excluded items
 		a.OrderedList = nil
-		a.OptionalSet = nil
+		a.OptionalSet = new(IntSet) // todo can replace with clear?
 	}
 	if anyOrder {
 		// allow any order
@@ -431,7 +432,7 @@ func (a IntSignature) String() string {
 		// element ordering is strict
 		list = a.OrderedList
 	} else {
-		if a.OptionalSet == nil {
+		if a.RequiredSet.Len() == 0 {
 			buf.WriteByte(flagAnyItems)
 		} else {
 			buf.WriteByte(flagAnyOrder)
@@ -449,11 +450,11 @@ func (a IntSignature) String() string {
 			buf.WriteString(fieldElemSep)
 		}
 		switch {
-		case a.OptionalSet[elem]:
+		case a.OptionalSet.Has(elem):
 			buf.WriteByte(flagOptional)
-		case a.UnlikelySet[elem]:
+		case a.UnlikelySet.Has(elem):
 			buf.WriteByte(flagUnlikely)
-		case a.ExcludedSet[elem]:
+		case a.ExcludedSet.Has(elem):
 			buf.WriteByte(flagExcluded)
 		}
 		buf.WriteString(fmt.Sprintf("%x", elem))
@@ -542,35 +543,58 @@ func (a IntSignature) Merge(b IntSignature) (merged IntSignature) {
 	// 2) The order of elements in a and b must remain the same.
 	// 3) If there exists elements e1, e2 that appear in different orders
 	// in a and b, the merged list should be nil (accept any ordering).
+	fmt.Println("BEFORE MERGE A'S REQUIRED SET IS", a.RequiredSet.String())
+	fmt.Println("BEFORE MERGE A'S OPTIONAL SET IS", a.OptionalSet.String())
+	fmt.Println("BEFORE MERGE A'S EXCLUDED SET IS", a.ExcludedSet.String())
+	fmt.Println("BEFORE MERGE A'S UNLIKELY SET IS", a.UnlikelySet.String())
+	fmt.Println("BEFORE MERGE A'S ORDERED LIST IS", a.OrderedList)
+
+	fmt.Println("BEFORE MERGE B'S REQUIRED SET IS", b.RequiredSet.String())
+	fmt.Println("BEFORE MERGE B'S OPTIONAL SET IS", b.OptionalSet.String())
+	fmt.Println("BEFORE MERGE B'S EXCLUDED SET IS", b.ExcludedSet.String())
+	fmt.Println("BEFORE MERGE B'S UNLIKELY SET IS", b.UnlikelySet.String())
+	fmt.Println("BEFORE MERGE B'S ORDERED LIST IS", b.OrderedList)
+
+	merged = IntSignature{
+		IntList{},
+		new(IntSet),
+		new(IntSet),
+		new(IntSet),
+		new(IntSet),
+	}
+
 	anyOrder := false
 	if a.OrderedList == nil || b.OrderedList == nil {
 		anyOrder = true
 	} else {
-		mergedSet := make(IntSet)
-		merged.OrderedList = IntList{}
-		bSet := b.RequiredSet.Union(b.OptionalSet).Union(b.UnlikelySet)
+		var mergedSet IntSet
+		var bSet IntSet
+		bSet.Copy(b.RequiredSet.Union(b.OptionalSet).Union(b.UnlikelySet))
 		bIdx := 0
 		bLen := len(b.OrderedList)
 		for _, elem := range a.OrderedList {
+			fmt.Println(elem)
 			// check if elem is already merged
-			if mergedSet[elem] {
+			if mergedSet.Has(elem) {
 				// elem is already merged, so abort and accept any ordering
+				fmt.Println("ANY ORDER TRUE")
 				anyOrder = true
 				break
 			}
 			// check if b contains elem
-			if bSet[elem] {
+			if bSet.Has(elem) {
+				fmt.Println("bSet has elem!!")
 				// add all elems of b up to elem
 				for ; bIdx < bLen && b.OrderedList[bIdx] != elem; bIdx++ {
 					merged.OrderedList = append(merged.OrderedList, b.OrderedList[bIdx])
-					mergedSet[b.OrderedList[bIdx]] = true
+					mergedSet.Insert(b.OrderedList[bIdx])
 				}
 				// skip past elem since it is added below
 				bIdx++
 			}
-			// add elem to merged list/set
+			// add elem to merged list and set
 			merged.OrderedList = append(merged.OrderedList, elem)
-			mergedSet[elem] = true
+			mergedSet.Insert(elem)
 		}
 		// add remaining elems of b to merged list
 		merged.OrderedList = append(merged.OrderedList, b.OrderedList[bIdx:bLen]...)
@@ -579,32 +603,46 @@ func (a IntSignature) Merge(b IntSignature) (merged IntSignature) {
 	// Clear ordered list if any ordering is accepted
 	if anyOrder {
 		merged.OrderedList = nil
+		fmt.Println(merged.OrderedList)
 	}
 
 	// Take intersection of required elems
-	if a.RequiredSet != nil || b.RequiredSet != nil {
-		merged.RequiredSet = a.RequiredSet.Inter(b.RequiredSet)
+	if a.RequiredSet.Len() != 0 || b.RequiredSet.Len() != 0 {
+		merged.RequiredSet.Copy(a.RequiredSet.Inter(b.RequiredSet))
 	}
 
 	// Take intersection of excluded elems
-	if a.ExcludedSet != nil || b.ExcludedSet != nil {
-		merged.ExcludedSet = a.ExcludedSet.Inter(b.ExcludedSet)
+	if a.ExcludedSet.Len() != 0 || b.ExcludedSet.Len() != 0 {
+		merged.ExcludedSet.Copy(a.ExcludedSet.Inter(b.ExcludedSet))
 	}
 
 	// Take union of optional elems
-	if a.OptionalSet == nil || b.OptionalSet == nil {
-		merged.OptionalSet = nil
-	} else {
-		merged.OptionalSet = a.OptionalSet.Union(b.OptionalSet).Union(a.RequiredSet).Union(b.RequiredSet).Diff(merged.RequiredSet)
+	//if anyOrder && (a.OptionalSet.Len() == 0 || b.OptionalSet.Len() == 0) {
+	//	merged.OptionalSet = new(IntSet)
+	//} else {
+	//	merged.OptionalSet.Copy(a.OptionalSet.Union(b.OptionalSet).Union(a.RequiredSet).Union(b.RequiredSet).Diff(merged.RequiredSet))
+	//}
+	// todo... think about new optional set merging rules
+	if a.RequiredSet.Len() != 0 && b.RequiredSet.Len() != 0 {
+		merged.OptionalSet.Copy(a.OptionalSet.Union(b.OptionalSet).Union(a.RequiredSet).Union(b.RequiredSet).Diff(merged.RequiredSet))
 	}
 
 	// Take union of unlikely elems
-	if a.UnlikelySet == nil || b.UnlikelySet == nil {
-		merged.UnlikelySet = nil
-	} else {
-		merged.UnlikelySet = a.UnlikelySet.Union(b.UnlikelySet).Union(a.OptionalSet).Union(b.OptionalSet).Diff(merged.OptionalSet)
+	//if a.UnlikelySet.Len() == 0 || b.UnlikelySet.Len() == 0 {
+	//	merged.UnlikelySet = new(IntSet)
+	//} else {
+	//	merged.UnlikelySet.Copy(a.UnlikelySet.Union(b.UnlikelySet).Union(a.OptionalSet).Union(b.OptionalSet).Diff(merged.OptionalSet))
+	//}
+	// todo... think about new unlikely set merging rules
+	if a.OptionalSet.Len() != 0 && b.OptionalSet.Len() != 0 {
+		merged.UnlikelySet.Copy(a.UnlikelySet.Union(b.UnlikelySet).Union(a.OptionalSet).Union(b.OptionalSet).Diff(merged.OptionalSet))
 	}
 
+	fmt.Println("AFTER MERGE REQUIRED SET IS", merged.RequiredSet.String())
+	fmt.Println("AFTER MERGE OPTIONAL SET IS", merged.OptionalSet.String())
+	fmt.Println("AFTER MERGE EXCLUDED SET IS", merged.ExcludedSet.String())
+	fmt.Println("AFTER MERGE UNLIKELY SET IS", merged.UnlikelySet.String())
+	fmt.Println("AFTER MERGE ORDERED LIST IS", merged.OrderedList)
 	return
 }
 
@@ -781,24 +819,24 @@ func (a StringSignature) Match(list StringList) Match {
 		return MatchImpossible
 	}
 	// check that the set does not contain any excluded items
-	if len(set.Inter(a.ExcludedSet)) > 0 {
+	if set.Inter(a.ExcludedSet).Len() > 0 {
 		return MatchImpossible
 	}
 	// check that the set has all required items
-	if len(a.RequiredSet.Diff(set)) > 0 {
+	if a.RequiredSet.Diff(set).Len() > 0 {
 		return MatchImpossible
 	}
 	// see if there's anything left after removing required and optional items
-	set = set.Diff(a.RequiredSet).Diff(a.OptionalSet)
-	if a.OptionalSet != nil && len(set) > 0 {
+	set.Copy(set.Diff(a.RequiredSet).Diff(a.OptionalSet))
+	if a.OptionalSet.Len() != 0 && set.Len() > 0 {
 		// check if the remaining items are unlikely or impossible
-		if a.UnlikelySet != nil && len(set.Diff(a.UnlikelySet)) > 0 {
+		if a.UnlikelySet.Len() != 0 && set.Diff(a.UnlikelySet).Len() > 0 {
 			return MatchImpossible
 		}
 		return MatchUnlikely
 	}
 	// check if the set has any unlikely items
-	if len(set.Inter(a.UnlikelySet)) > 0 {
+	if set.Inter(a.UnlikelySet).Len() > 0 {
 		return MatchUnlikely
 	}
 	return MatchPossible