This PR introduces use of [intsets](https://godoc.org/golang.org/x/tools/container/intsets), a golang package that implements a fast set implementation for sparse ints. 

Repeated trials on my local dev environment demonstrate that performance improvement is substantial (a 38% decrease from 19038614 ns/op to 11796146 ns/op), and there is also a decrease in memory usage (17% decrease from 1070.14MB to 891.50MB). These stats are included below.

Old master branch cpu and mem pprof output:
```
BenchmarkProcessorCheck-8   	     100	  19038614 ns/op
PASS
ok  	github.com/cloudflare/mitmengine	2.582s
Gabbis-Work-MacBook-3:mitmengine gabbifisher$ go tool pprof -top cpu.prof 
Type: cpu
Time: Dec 9, 2018 at 1:47pm (PST)
Duration: 2.56s, Total samples = 3.07s (119.80%)
Showing nodes accounting for 2.96s, 96.42% of 3.07s total
Dropped 54 nodes (cum <= 0.02s)
      flat  flat%   sum%        cum   cum%
     1.21s 39.41% 39.41%      1.21s 39.41%  syscall.Syscall
     0.53s 17.26% 56.68%      0.53s 17.26%  runtime.pthread_cond_wait
     0.15s  4.89% 61.56%      0.15s  4.89%  runtime.stkbucket
     0.14s  4.56% 66.12%      0.14s  4.56%  runtime.duffcopy
     0.13s  4.23% 70.36%      0.13s  4.23%  runtime.kevent
     0.11s  3.58% 73.94%      0.11s  3.58%  runtime.pthread_cond_signal
     0.07s  2.28% 76.22%      0.07s  2.28%  runtime.pthread_cond_timedwait_relative_np
     0.04s  1.30% 77.52%      0.07s  2.28%  runtime.findObject
     0.04s  1.30% 78.83%      0.05s  1.63%  runtime.greyobject
     0.04s  1.30% 80.13%      0.05s  1.63%  runtime.heapBitsSetType
     0.04s  1.30% 81.43%      0.31s 10.10%  runtime.mallocgc
     0.04s  1.30% 82.74%      0.04s  1.30%  runtime.mapiternext
     0.03s  0.98% 83.71%      0.03s  0.98%  runtime.mapaccess1_fast64
     0.03s  0.98% 84.69%      0.03s  0.98%  runtime.memmove
     0.03s  0.98% 85.67%      0.15s  4.89%  runtime.scanobject
     0.03s  0.98% 86.64%      0.03s  0.98%  runtime.usleep
     0.02s  0.65% 87.30%      0.40s 13.03%  github.com/cloudflare/mitmengine/db.Database.GetBy
     0.02s  0.65% 87.95%      0.02s  0.65%  runtime.(*gcBitsArena).tryAlloc (inline)
     0.02s  0.65% 88.60%      0.02s  0.65%  runtime.add (inline)
     0.02s  0.65% 89.25%      0.02s  0.65%  runtime.arenaIndex (inline)
     0.02s  0.65% 89.90%      0.03s  0.98%  runtime.lock
     0.02s  0.65% 90.55%      0.02s  0.65%  runtime.memclrNoHeapPointers
     0.02s  0.65% 91.21%      0.22s  7.17%  runtime.newobject
     0.02s  0.65% 91.86%      0.02s  0.65%  runtime.nextFreeFast (inline)
     0.02s  0.65% 92.51%      0.03s  0.98%  runtime.spanOf (inline)
     0.02s  0.65% 93.16%      0.03s  0.98%  runtime.sweepone
     0.01s  0.33% 93.49%      1.28s 41.69%  github.com/cloudflare/mitmengine/db.(*Database).Load
     0.01s  0.33% 93.81%      0.12s  3.91%  github.com/cloudflare/mitmengine/fputil.(*IntSignature).Parse
     0.01s  0.33% 94.14%      0.05s  1.63%  github.com/cloudflare/mitmengine/fputil.StringSet.Diff
     0.01s  0.33% 94.46%      0.03s  0.98%  runtime.(*mcentral).grow
     0.01s  0.33% 94.79%      0.02s  0.65%  runtime.casgstatus
     0.01s  0.33% 95.11%      0.03s  0.98%  runtime.evacuate_fast64
     0.01s  0.33% 95.44%      0.23s  7.49%  runtime.gcDrain
     0.01s  0.33% 95.77%      0.07s  2.28%  runtime.mapassign_fast64
     0.01s  0.33% 96.09%      0.02s  0.65%  runtime.mapiterinit
     0.01s  0.33% 96.42%      0.02s  0.65%  runtime.wbBufFlush1
         0     0% 96.42%      1.07s 34.85%  bufio.(*Scanner).Scan
         0     0% 96.42%      0.21s  6.84%  github.com/cloudflare/mitmengine.(*Processor).Check
         0     0% 96.42%      1.40s 45.60%  github.com/cloudflare/mitmengine.(*Processor).Load
         0     0% 96.42%      0.14s  4.56%  github.com/cloudflare/mitmengine.LoadFile
         0     0% 96.42%      1.40s 45.60%  github.com/cloudflare/mitmengine.NewProcessor
         0     0% 96.42%      0.19s  6.19%  github.com/cloudflare/mitmengine/db.(*Record).Parse
         0     0% 96.42%      0.22s  7.17%  github.com/cloudflare/mitmengine/db.Database.GetByRequestFingerprint
         0     0% 96.42%      0.21s  6.84%  github.com/cloudflare/mitmengine/db.Database.GetByRequestFingerprint.func1
         0     0% 96.42%      0.18s  5.86%  github.com/cloudflare/mitmengine/db.Database.GetByUAFingerprint
         0     0% 96.42%      0.06s  1.95%  github.com/cloudflare/mitmengine/db.Database.GetByUAFingerprint.func1
         0     0% 96.42%      1.28s 41.69%  github.com/cloudflare/mitmengine/db.NewDatabase
         0     0% 96.42%      0.13s  4.23%  github.com/cloudflare/mitmengine/fputil.(*RequestSignature).Parse
         0     0% 96.42%      0.03s  0.98%  github.com/cloudflare/mitmengine/fputil.(*StringSignature).Parse
         0     0% 96.42%      0.04s  1.30%  github.com/cloudflare/mitmengine/fputil.(*UASignature).Parse
         0     0% 96.42%      0.03s  0.98%  github.com/cloudflare/mitmengine/fputil.IntList.Set
         0     0% 96.42%      0.15s  4.89%  github.com/cloudflare/mitmengine/fputil.IntSet.Inter
         0     0% 96.42%      0.19s  6.19%  github.com/cloudflare/mitmengine/fputil.IntSignature.Match
         0     0% 96.42%      0.25s  8.14%  github.com/cloudflare/mitmengine/fputil.RequestSignature.Match
         0     0% 96.42%      0.26s  8.47%  github.com/cloudflare/mitmengine/fputil.RequestSignature.MatchMap
         0     0% 96.42%      0.06s  1.95%  github.com/cloudflare/mitmengine/fputil.StringSignature.Match
         0     0% 96.42%      0.02s  0.65%  github.com/cloudflare/mitmengine/fputil.UASignature.Match
         0     0% 96.42%      1.57s 51.14%  github.com/cloudflare/mitmengine_test.BenchmarkProcessorCheck
         0     0% 96.42%      0.02s  0.65%  github.com/cloudflare/mitmengine_test.TestProcessorConfigFile.func1
         0     0% 96.42%      0.02s  0.65%  github.com/cloudflare/mitmengine_test.TestProcessorConfigFile.func2
         0     0% 96.42%      0.17s  5.54%  github.com/cloudflare/mitmengine_test.TestProcessorConfigFile.func3
         0     0% 96.42%      0.11s  3.58%  github.com/cloudflare/mitmengine_test.TestProcessorConfigFile.func4
         0     0% 96.42%      1.59s 51.79%  github.com/cloudflare/mitmengine_test._TestProcessorCheck
         0     0% 96.42%      0.11s  3.58%  github.com/cloudflare/mitmengine_test._TestProcessorGetByRequestSignatureMitm
         0     0% 96.42%      0.17s  5.54%  github.com/cloudflare/mitmengine_test._TestProcessorGetByUASignatureBrowser
         0     0% 96.42%      1.07s 34.85%  internal/poll.(*FD).Read
         0     0% 96.42%      1.07s 34.85%  os.(*File).Read
         0     0% 96.42%      1.07s 34.85%  os.(*File).read
         0     0% 96.42%      0.14s  4.56%  os.Open
         0     0% 96.42%      0.14s  4.56%  os.OpenFile
         0     0% 96.42%      0.14s  4.56%  os.openFileNolog
         0     0% 96.42%      0.03s  0.98%  runtime.(*gcWork).balance
         0     0% 96.42%      0.03s  0.98%  runtime.(*mcache).nextFree
         0     0% 96.42%      0.03s  0.98%  runtime.(*mcache).nextFree.func1
         0     0% 96.42%      0.03s  0.98%  runtime.(*mcache).refill
         0     0% 96.42%      0.03s  0.98%  runtime.(*mcentral).cacheSpan
         0     0% 96.42%      0.03s  0.98%  runtime.bgsweep
         0     0% 96.42%      0.02s  0.65%  runtime.execute
         0     0% 96.42%      0.55s 17.92%  runtime.findrunnable
         0     0% 96.42%      0.06s  1.95%  runtime.forEachP
         0     0% 96.42%      0.22s  7.17%  runtime.gcBgMarkWorker
         0     0% 96.42%      0.22s  7.17%  runtime.gcBgMarkWorker.func2
         0     0% 96.42%      0.06s  1.95%  runtime.gcMarkDone.func1
         0     0% 96.42%      0.13s  4.23%  runtime.gcStart.func2
         0     0% 96.42%      0.02s  0.65%  runtime.gcWriteBarrier
         0     0% 96.42%      0.02s  0.65%  runtime.gentraceback
         0     0% 96.42%      0.08s  2.61%  runtime.gopreempt_m
         0     0% 96.42%      0.12s  3.91%  runtime.goschedImpl
         0     0% 96.42%      0.04s  1.30%  runtime.gosched_m
         0     0% 96.42%      0.03s  0.98%  runtime.gosweepone
         0     0% 96.42%      0.03s  0.98%  runtime.gosweepone.func1
         0     0% 96.42%      0.03s  0.98%  runtime.growWork_fast64
         0     0% 96.42%      0.02s  0.65%  runtime.growslice
         0     0% 96.42%      0.03s  0.98%  runtime.handoff
         0     0% 96.42%      0.16s  5.21%  runtime.mProf_Malloc
         0     0% 96.42%      0.05s  1.63%  runtime.makeBucketArray
         0     0% 96.42%      0.13s  4.23%  runtime.makemap
         0     0% 96.42%      0.09s  2.93%  runtime.makemap_small
         0     0% 96.42%      0.04s  1.30%  runtime.makeslice
         0     0% 96.42%      0.02s  0.65%  runtime.markroot
         0     0% 96.42%      0.63s 20.52%  runtime.mcall
         0     0% 96.42%      0.08s  2.61%  runtime.morestack
         0     0% 96.42%      0.22s  7.17%  runtime.mstart
         0     0% 96.42%      0.13s  4.23%  runtime.netpoll
         0     0% 96.42%      0.02s  0.65%  runtime.newMarkBits
         0     0% 96.42%      0.05s  1.63%  runtime.newarray
         0     0% 96.42%      0.08s  2.61%  runtime.newstack
         0     0% 96.42%      0.53s 17.26%  runtime.notesleep
         0     0% 96.42%      0.07s  2.28%  runtime.notetsleep
         0     0% 96.42%      0.07s  2.28%  runtime.notetsleep_internal
         0     0% 96.42%      0.12s  3.91%  runtime.notewakeup
         0     0% 96.42%      0.02s  0.65%  runtime.osyield
         0     0% 96.42%      0.59s 19.22%  runtime.park_m
         0     0% 96.42%      0.16s  5.21%  runtime.profilealloc
         0     0% 96.42%      0.07s  2.28%  runtime.resetspinning
         0     0% 96.42%      0.05s  1.63%  runtime.runSafePointFn
         0     0% 96.42%      0.70s 22.80%  runtime.schedule
         0     0% 96.42%      0.60s 19.54%  runtime.semasleep
         0     0% 96.42%      0.12s  3.91%  runtime.semawakeup
         0     0% 96.42%      0.13s  4.23%  runtime.startTheWorldWithSema
         0     0% 96.42%      0.07s  2.28%  runtime.startm
         0     0% 96.42%      0.54s 17.59%  runtime.stopm
         0     0% 96.42%      0.51s 16.61%  runtime.systemstack
         0     0% 96.42%      0.07s  2.28%  runtime.wakep
         0     0% 96.42%      0.02s  0.65%  runtime.wbBufFlush
         0     0% 96.42%      0.02s  0.65%  runtime.wbBufFlush.func1
         0     0% 96.42%      0.05s  1.63%  strings.Split
         0     0% 96.42%      0.05s  1.63%  strings.genSplit
         0     0% 96.42%      0.13s  4.23%  syscall.Open
         0     0% 96.42%      1.07s 34.85%  syscall.Read
         0     0% 96.42%      1.07s 34.85%  syscall.read
         0     0% 96.42%      1.55s 50.49%  testing.(*B).launch
         0     0% 96.42%      0.02s  0.65%  testing.(*B).run1.func1
         0     0% 96.42%      1.57s 51.14%  testing.(*B).runN
         0     0% 96.42%      0.32s 10.42%  testing.tRunner

Showing nodes accounting for 1058.10MB, 98.87% of 1070.14MB total
Dropped 39 nodes (cum <= 5.35MB)
      flat  flat%   sum%        cum   cum%
  467.15MB 43.65% 43.65%   573.20MB 53.56%  github.com/cloudflare/mitmengine/fputil.(*IntSignature).Parse
  167.56MB 15.66% 59.31%   167.56MB 15.66%  strings.genSplit
   82.50MB  7.71% 67.02%       83MB  7.76%  github.com/cloudflare/mitmengine/fputil.(*StringSignature).Parse
   77.02MB  7.20% 74.22%    77.02MB  7.20%  github.com/cloudflare/mitmengine/fputil.IntSet.Inter
   68.66MB  6.42% 80.63%    68.66MB  6.42%  github.com/cloudflare/mitmengine/db.(*Database).Add (inline)
   42.17MB  3.94% 84.57%    42.17MB  3.94%  regexp.(*bitState).reset
   39.01MB  3.65% 88.22%    39.01MB  3.65%  github.com/cloudflare/mitmengine/fputil.IntSet.Diff
   28.50MB  2.66% 90.88%    28.50MB  2.66%  github.com/cloudflare/mitmengine/fputil.StringSet.Diff
   28.01MB  2.62% 93.50%    28.01MB  2.62%  github.com/cloudflare/mitmengine/fputil.IntList.Set
   26.51MB  2.48% 95.98%    26.51MB  2.48%  bufio.(*Scanner).Text (inline)
      14MB  1.31% 97.29%       14MB  1.31%  github.com/cloudflare/mitmengine/fputil.StringSet.Inter
   10.50MB  0.98% 98.27%    10.50MB  0.98%  github.com/cloudflare/mitmengine/fputil.StringList.Set
    5.50MB  0.51% 98.78%   170.54MB 15.94%  github.com/cloudflare/mitmengine/fputil.RequestSignature.MatchMap
    0.50MB 0.047% 98.83%   809.38MB 75.63%  github.com/cloudflare/mitmengine.(*Processor).Load
    0.50MB 0.047% 98.87%   179.54MB 16.78%  github.com/cloudflare/mitmengine/db.Database.GetBy
         0     0% 98.87%   114.52MB 10.70%  github.com/cloudflare/mitmengine.(*Processor).Check
         0     0% 98.87%   809.38MB 75.63%  github.com/cloudflare/mitmengine.NewProcessor
         0     0% 98.87%   814.88MB 76.15%  github.com/cloudflare/mitmengine/db.(*Database).Load
         0     0% 98.87%   717.71MB 67.07%  github.com/cloudflare/mitmengine/db.(*Record).Parse
         0     0% 98.87%   147.54MB 13.79%  github.com/cloudflare/mitmengine/db.Database.GetByRequestFingerprint
         0     0% 98.87%   147.54MB 13.79%  github.com/cloudflare/mitmengine/db.Database.GetByRequestFingerprint.func1
         0     0% 98.87%       32MB  2.99%  github.com/cloudflare/mitmengine/db.Database.GetByUAFingerprint
         0     0% 98.87%    31.50MB  2.94%  github.com/cloudflare/mitmengine/db.Database.GetByUAFingerprint.func1
         0     0% 98.87%   814.88MB 76.15%  github.com/cloudflare/mitmengine/db.NewDatabase
         0     0% 98.87%     6.50MB  0.61%  github.com/cloudflare/mitmengine/fputil.(*MitmInfo).Parse
         0     0% 98.87%   640.71MB 59.87%  github.com/cloudflare/mitmengine/fputil.(*RequestSignature).Parse
         0     0% 98.87%       62MB  5.79%  github.com/cloudflare/mitmengine/fputil.(*UASignature).Parse
         0     0% 98.87%     6.50MB  0.61%  github.com/cloudflare/mitmengine/fputil.(*UAVersion).Parse
         0     0% 98.87%    14.50MB  1.35%  github.com/cloudflare/mitmengine/fputil.(*UAVersionSignature).Parse
         0     0% 98.87%   144.04MB 13.46%  github.com/cloudflare/mitmengine/fputil.IntSignature.Match
         0     0% 98.87%   169.54MB 15.84%  github.com/cloudflare/mitmengine/fputil.RequestSignature.Match
         0     0% 98.87%    52.50MB  4.91%  github.com/cloudflare/mitmengine/fputil.StringSignature.Match
         0     0% 98.87%    31.50MB  2.94%  github.com/cloudflare/mitmengine/fputil.UASignature.Match
         0     0% 98.87%   912.40MB 85.26%  github.com/cloudflare/mitmengine_test.BenchmarkProcessorCheck
         0     0% 98.87%     7.50MB   0.7%  github.com/cloudflare/mitmengine_test.TestProcessorConfigFile.func2
         0     0% 98.87%    65.56MB  6.13%  github.com/cloudflare/mitmengine_test.TestProcessorConfigFile.func3
         0     0% 98.87%    78.14MB  7.30%  github.com/cloudflare/mitmengine_test.TestProcessorConfigFile.func4
         0     0% 98.87%   919.91MB 85.96%  github.com/cloudflare/mitmengine_test._TestProcessorCheck
         0     0% 98.87%    78.14MB  7.30%  github.com/cloudflare/mitmengine_test._TestProcessorGetByRequestSignatureMitm
         0     0% 98.87%    65.56MB  6.13%  github.com/cloudflare/mitmengine_test._TestProcessorGetByUASignatureBrowser
         0     0% 98.87%    40.56MB  3.79%  github.com/cloudflare/mitmengine_test.uaSigToFin
         0     0% 98.87%    42.67MB  3.99%  regexp.(*Regexp).ReplaceAllString
         0     0% 98.87%    42.67MB  3.99%  regexp.(*Regexp).doExecute
         0     0% 98.87%    42.67MB  3.99%  regexp.(*Regexp).replaceAll
         0     0% 98.87%    42.17MB  3.94%  regexp.(*machine).backtrack
         0     0% 98.87%   159.56MB 14.91%  strings.Split
         0     0% 98.87%        8MB  0.75%  strings.SplitN
         0     0% 98.87%   902.90MB 84.37%  testing.(*B).launch
         0     0% 98.87%     9.50MB  0.89%  testing.(*B).run1.func1
         0     0% 98.87%   912.40MB 85.26%  testing.(*B).runN
         0     0% 98.87%   155.70MB 14.55%  testing.tRunner
```

New intset implementation pprof output!
```
Gabbis-Work-MacBook-3:mitmengine gabbifisher$ go test -cpuprofile cpu.prof -memprofile mem.prof -bench .
2018/12/09 21:31:45 WARNING: loading file "" produced error "open : no such file or directory"
2018/12/09 21:31:45 WARNING: loading file "" produced error "open : no such file or directory"
2018/12/09 21:31:45 WARNING: loading file "" produced error "open : no such file or directory"
goos: darwin
goarch: amd64
pkg: github.com/cloudflare/mitmengine
BenchmarkProcessorCheck-8   	     100	  11796146 ns/op
PASS
ok  	github.com/cloudflare/mitmengine	1.648s
Gabbis-Work-MacBook-3:mitmengine gabbifisher$ go tool pprof -top cpu.prof 
Type: cpu
Time: Dec 9, 2018 at 9:31pm (PST)
Duration: 1.63s, Total samples = 2.12s (129.95%)
Showing nodes accounting for 2.05s, 96.70% of 2.12s total
Dropped 44 nodes (cum <= 0.01s)
      flat  flat%   sum%        cum   cum%
     1.04s 49.06% 49.06%      1.04s 49.06%  syscall.Syscall
     0.34s 16.04% 65.09%      0.34s 16.04%  runtime.pthread_cond_wait
     0.16s  7.55% 72.64%      0.16s  7.55%  runtime.pthread_cond_signal
     0.11s  5.19% 77.83%      0.11s  5.19%  runtime.kevent
     0.09s  4.25% 82.08%      0.09s  4.25%  runtime.stkbucket
     0.08s  3.77% 85.85%      0.08s  3.77%  runtime.duffcopy
     0.07s  3.30% 89.15%      0.07s  3.30%  runtime.pthread_cond_timedwait_relative_np
     0.04s  1.89% 91.04%      0.04s  1.89%  runtime.findObject
     0.03s  1.42% 92.45%      0.03s  1.42%  runtime.mapiternext
     0.02s  0.94% 93.40%      0.02s  0.94%  runtime.memclrNoHeapPointers
     0.02s  0.94% 94.34%      0.02s  0.94%  runtime.procyield
     0.02s  0.94% 95.28%      0.02s  0.94%  strings.genSplit
     0.01s  0.47% 95.75%      0.06s  2.83%  github.com/cloudflare/mitmengine/fputil.(*IntSignature).Parse
     0.01s  0.47% 96.23%      0.06s  2.83%  github.com/cloudflare/mitmengine/fputil.RequestSignature.Match
     0.01s  0.47% 96.70%      0.02s  0.94%  runtime.gcAssistAlloc1
         0     0% 96.70%      0.88s 41.51%  bufio.(*Scanner).Scan
         0     0% 96.70%      0.06s  2.83%  github.com/cloudflare/mitmengine.(*Processor).Check
         0     0% 96.70%      1.11s 52.36%  github.com/cloudflare/mitmengine.(*Processor).Load
         0     0% 96.70%      0.16s  7.55%  github.com/cloudflare/mitmengine.LoadFile
         0     0% 96.70%      1.11s 52.36%  github.com/cloudflare/mitmengine.NewProcessor
         0     0% 96.70%      0.96s 45.28%  github.com/cloudflare/mitmengine/db.(*Database).Load
         0     0% 96.70%      0.08s  3.77%  github.com/cloudflare/mitmengine/db.(*Record).Parse
         0     0% 96.70%      0.18s  8.49%  github.com/cloudflare/mitmengine/db.Database.GetBy
         0     0% 96.70%      0.05s  2.36%  github.com/cloudflare/mitmengine/db.Database.GetByRequestFingerprint
         0     0% 96.70%      0.05s  2.36%  github.com/cloudflare/mitmengine/db.Database.GetByRequestFingerprint.func1
         0     0% 96.70%      0.13s  6.13%  github.com/cloudflare/mitmengine/db.Database.GetByUAFingerprint
         0     0% 96.70%      0.03s  1.42%  github.com/cloudflare/mitmengine/db.Database.GetByUAFingerprint.func1
         0     0% 96.70%      0.96s 45.28%  github.com/cloudflare/mitmengine/db.NewDatabase
         0     0% 96.70%      0.03s  1.42%  github.com/cloudflare/mitmengine/fputil.(*IntSet).Inter
         0     0% 96.70%      0.06s  2.83%  github.com/cloudflare/mitmengine/fputil.(*RequestSignature).Parse
         0     0% 96.70%      0.02s  0.94%  github.com/cloudflare/mitmengine/fputil.(*StringSignature).Parse
         0     0% 96.70%      0.02s  0.94%  github.com/cloudflare/mitmengine/fputil.(*UASignature).Parse
         0     0% 96.70%      0.05s  2.36%  github.com/cloudflare/mitmengine/fputil.IntSignature.Match
         0     0% 96.70%      0.05s  2.36%  github.com/cloudflare/mitmengine/fputil.RequestSignature.MatchMap
…